refactor: implement Tor daemon management for anonymous metrics forwarding
