+++ packages/dappmanager/src/index.ts
@@ -26,63 +26,9 @@ import { DeviceCalls } from "./calls/device/index.js";
 import { startHttpApi } from "./api/startHttpApi.js";
 import { DappNodeDirectory, DappNodeRegistry, MultiUrlJsonRpcProvider } from "@dappnode/toolkit";
 import { Consensus, Execution, MevBoost, Signer } from "@dappnode/stakers";
-import { spawn } from "child_process";
+import { startTor, stopTor } from "./tor/index.js";
 
 const controller = new AbortController();
-
-// Tor process reference for cleanup
-let torProcess: ReturnType<typeof spawn> | null = null;
-
-/**
- * Start Tor daemon for anonymous metrics forwarding.
- * Spawns Tor as a background process and waits for SOCKS5 proxy to be ready.
- */
-async function startTor(): Promise<void> {
-  return new Promise((resolve, reject) => {
-    logs.info("Starting Tor daemon for anonymous metrics...");
-
-    torProcess = spawn("tor", [], {
-      stdio: ["ignore", "pipe", "pipe"],
-      detached: false
-    });
-
-    let started = false;
-
-    torProcess.stdout?.on("data", (data: Buffer) => {
-      const output = data.toString();
-      // Tor logs "Bootstrapped 100%" when fully ready
-      if (output.includes("Bootstrapped 100%") && !started) {
-        started = true;
-        logs.info("Tor daemon started successfully (SOCKS5 proxy on 127.0.0.1:9050)");
-        resolve();
-      }
-    });
-
-    torProcess.stderr?.on("data", (data: Buffer) => {
-      logs.warn(`Tor stderr: ${data.toString().trim()}`);
-    });
-
-    torProcess.on("error", (err) => {
-      logs.error("Failed to start Tor daemon", err);
-      if (!started) reject(err);
-    });
-
-    torProcess.on("exit", (code) => {
-      if (code !== 0 && code !== null) {
-        logs.warn(`Tor daemon exited with code ${code}`);
-      }
-      torProcess = null;
-    });
-
-    // Timeout after 60 seconds if Tor doesn't start
-    setTimeout(() => {
-      if (!started) {
-        logs.warn("Tor daemon startup timeout, continuing without waiting for full bootstrap");
-        resolve();
-      }
-    }, 60000);
-  });
-}
 
 process.on("unhandledRejection", (reason) => {
   console.error("Unhandled Rejection:", reason);
@@ -96,11 +42,8 @@ process.on("SIGINT", () => {
   controller.abort();
   server.close();
   uiMetricsServer.close();
-  // Kill Tor process if running
-  if (torProcess) {
-    torProcess.kill();
-    torProcess = null;
-  }
+  // Stop Tor daemon if running
+  stopTor();
   process.exit(0);
 });
