--- packages/dappmanager/src/utils/README.md
@@ -1,3 +1,0 @@
-# Utils
-
-A util can only depend on other utils or external npm packages.
